# src/http-server/openapi/restqa/restqa.yml
openapi: 3.0.0
info:
  version: 0.0.1
  contact:
    name: RestQA support
    email: hello@mailqa.io
  license:
    name: RestQA
    url: 'https://restqa.io'
  title: RestQA 🦏
  description: |
    ## Description
    
    RestQA also provide a way to control your framework through a set of RestQA API.
    
    ### Why ?
    
    Actually some project are very specific and could face issue to run the test close to the target environment through Continuous integration.
    But maybe we can think of a project that is on premise and the best way to test this api is to have an access from inside. Then you can just deploy RestQA on the same server and start playing with the api 🚀
    
    ### Installation.
    
    You only need 3 steps to access to theses apis:
    
    1. Install RestQA `npm i -g @restqa/restqa`
    2. Initiate a RestQA project `restqa init`
    3. Run the RestQA server  `restqa serve`
    
    
    From there you should be able to access RestQA from it default port 8081 (https://localhot:8081)

    If you face any issue feel free to connect with us on [Discord](https://restqa.io/chat)
servers:
  - url: 'http://localhost:8081'
    description: The Dashboard API host
    variables:
      environment:
        enum:
          - local
        default: local
...
paths:
...
  /api/restqa/initialize:
    post:
      tags:
        - RestQA
      deprecated: false
      operationId: initialize
      summary: Initialize a RestQA project
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - env
                - url
              properties:
                name:
                  description: The project name
                  example: "Backend api"
                  type: "string"
                description:
                  description: The project description
                  example: "All the API used by the different frontends"
                  type: "string"
                env:
                  description: Define the first environement required to be setup
                  example: "local"
                  type: "string"
                url:
                  description: The host url of the api that required to be tested
                  example: "https://api.example.com"
                  type: "string"
                ci:
                  description: Define the continuous integration setup wanted
                  example: "github-action"
                  type: "string"
                  enum:
                    - 'github-action'
                    - 'gitlab-ci'
                    - 'bitbucket-pipeline'
                    - 'circle-ci'
                    - 'travis'
                    - 'jenkins'
                folder:
                  description: Define the location where the project needs to be initialized
                  example: "./test-integrations"
                  type: "string"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  configuration:
                    type: string
                    description: 'Location of the restqa configuration file'
                    example: '/home/john/project/.restqa.yml'
                  folder:
                    type: string
                    description: 'Location of the restqa project folder'
                    example: '/home/john/project'
        '406':
          description: Processing Error
          content:
            application/json:
              schema:
                $ref: '#/definitions/error406'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/definitions/error500'
...
  /reports:
    get:
      tags:
        - Reports
      deprecated: false
      operationId: GetReportList
      summary: Retrieve a list existing report on the server
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/definitions/report'
        '500':
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/definitions/error500'
    post:
      tags:
        - Reports
      deprecated: false
      operationId: CreateReport
      summary: Create an html report from a RestQA Json test result
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/example/reportJSON'

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/definitions/report'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/definitions/error500'
...
definitions:
...
  error406:
    type: object
    properties:
      message:
        type: string
        example: "The configuration file has an issue"
  error500:
    type: object
    properties:
      message:
        type: string
        example: "Internal server error"
...
  report:
    type: object
    properties:
      id:
        type: string
        example: 27e2d826-a2fd-4f01-abec-8471333112a5
        description: 'The Id of the suite report'
      url:
        type: string
        example: 'https://exmaple.com/reports/27e2d826-a2fd-4f01-abec-8471333112a5'
        description: 'The url of the generated html report'
...